Make the view work
