+++ compass/src/main/AndroidManifest.xml
@@ -1,11 +1,23 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.kmadsen.compass" >
+<?xml version="1.0" encoding="utf-8"?>
+<manifest
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.kmadsen.compass">
 
-    <uses-feature android:glEsVersion="0x00020000"/>
+    <uses-feature android:glEsVersion="0x00020000" />
 
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
     <uses-permission android:name="android.permission.CHANGE_WIFI_STATE" />
     <uses-permission android:name="android.permission.CHANGE_WIFI_MULTICAST_STATE" />
     <uses-permission android:name="android.permission.ACTIVITY_RECOGNITION" />
+
+    <application>
+
+        <activity
+            android:name=".sensors.config.SensorConfigActivity"
+            android:label="@string/compass_configure_sensors"
+            android:theme="@style/AppTheme.NoActionBar" />
+
+    </application>
+
 </manifest>
