+++ compass/src/main/java/com/kmadsen/compass/CompassMainActivity.kt
@@ -1,5 +1,6 @@
 package com.kmadsen.compass
 
+import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
@@ -8,7 +9,7 @@ import androidx.fragment.app.FragmentTransaction
 import com.kmadsen.compass.location.LocationSensor
 import com.kmadsen.compass.mapbox.MapViewController
 import com.kmadsen.compass.sensors.SensorLogger
-import com.kmadsen.compass.sensors.config.SensorConfigFragment
+import com.kmadsen.compass.sensors.config.SensorConfigActivity
 import com.kylemadsen.core.koin.koinLateModule
 import com.mapbox.mapboxsdk.maps.MapView
 import com.mapbox.mapboxsdk.maps.Style
@@ -61,6 +62,8 @@ class CompassMainActivity : AppCompatActivity() {
         }
 
         findViewById<Button>(R.id.configure_sensors).setOnClickListener {
+            val intent = Intent(this, SensorConfigActivity::class.java)
+            startActivity(intent)
         }
     }
